+++ src/tree/GitRepoItem.ts
@@ -7,12 +7,13 @@ export class GitRepoItem extends vscode.TreeItem {
     public readonly repoPath: string,
     public readonly branch: string,
     public readonly hasChanges: boolean,
+    public readonly localBranchCount: number,
     public readonly repo: SimpleGit,
   ) {
     super(path.basename(repoPath), vscode.TreeItemCollapsibleState.None);
 
-    this.tooltip = `${repoPath}\nГілка: ${branch}`;
-    this.description = branch;
+    this.tooltip = `${repoPath}\nГілка: ${branch} \n${localBranchCount} локальних гілок`;
+    this.description = `${branch} [${localBranchCount}]`;
 
     this.iconPath = new vscode.ThemeIcon(
       hasChanges ? 'warning' : 'check',